+++ elements/stencil.config.ts
@@ -15,6 +15,10 @@ export const config: Config = {
           src: join(__dirname, '../node_modules/@shoelace-style/shoelace/dist/assets'),
           dest: 'shoelace/assets',
         },
+        {
+          src: join(__dirname, '../node_modules/@uvdsl/solid-oidc-client-browser/dist/esm/web/RefreshWorker.js'),
+          dest: 'RefreshWorker.js',
+        },
       ],
     },
     {
@@ -33,6 +37,10 @@ export const config: Config = {
         {
           src: join(__dirname, '../node_modules/@shoelace-style/shoelace/dist/assets'),
           dest: 'build/shoelace/assets',
+        },
+        {
+          src: join(__dirname, '../node_modules/@uvdsl/solid-oidc-client-browser/dist/esm/web/RefreshWorker.js'),
+          dest: 'build/RefreshWorker.js',
         },
       ],
       serviceWorker: false, // disable stencils own service worker